requirements.txt
sodapy
kafka-python
pyspark
retrying
python-dotenv
pytest